% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace-ak.r
\docType{data}
\name{ak}
\alias{ak}
\title{Unified AK Namespace Facade}
\usage{
ak
}
\description{
\code{ak} is a lightweight facade that groups package exports by domain.
Use \code{ak$stats$...} for statistics-related APIs.
}
\keyword{datasets}
