% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-quant_matrix.r
\name{create_quant_matrix}
\alias{create_quant_matrix}
\title{Create a Quantification Matrix Object}
\usage{
create_quant_matrix(mat, col_label = NULL, row_label = NULL)
}
\arguments{
\item{mat}{A numeric-like matrix (or matrix-coercible object).}

\item{col_label}{Optional vector/list of column labels with length \code{ncol(mat)}.}

\item{row_label}{Optional vector/list of row labels with length \code{nrow(mat)}.}
}
\value{
A \code{QuantMatrix} object.
}
\description{
Create a Quantification Matrix Object
}
