% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-quant_matrix.r
\name{impute_knn}
\alias{impute_knn}
\title{KNN Impute a Quantification Matrix}
\usage{
impute_knn(x, k = 10)
}
\arguments{
\item{x}{A \code{QuantMatrix} object.}

\item{k}{Number of nearest neighbors to use for imputation.}
}
\value{
A new \code{QuantMatrix} object with imputed values.
}
\description{
Impute missing values by k-nearest neighbors.
}
