% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bar-faceted.r
\name{plot_bar_faceted}
\alias{plot_bar_faceted}
\title{Plot Faceted Bar Chart}
\usage{
plot_bar_faceted(df, col_x = "X", col_y = "Y", col_y_name = "YName")
}
\arguments{
\item{df}{A tabular object.}

\item{col_x}{Column name used for x-axis.}

\item{col_y}{Column name used for y-axis.}

\item{col_y_name}{Column name used for facet label.}
}
\value{
Currently returns \code{NULL} (placeholder implementation).
}
\description{
Plot Faceted Bar Chart
}
