% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-quant_matrix.r
\name{QuantMatrix}
\alias{QuantMatrix}
\title{Quantification Matrix Class}
\usage{
QuantMatrix(mat, col_label = NULL, row_label = NULL)
}
\description{
\code{QuantMatrix} stores the numeric matrix and optional per-axis labels.
}
