% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-quant_matrix.r
\name{QuantMatrix}
\alias{QuantMatrix}
\title{Quantification Matrix Class}
\usage{
QuantMatrix(mat, col_label = NULL, row_label = NULL)
}
\arguments{
\item{mat}{A numeric-like matrix (or matrix-coercible object).}

\item{col_label}{Optional vector/list of column labels
with length \code{ncol(mat)}.}

\item{row_label}{Optional vector/list of row labels with length \code{nrow(mat)}.}
}
\description{
\code{QuantMatrix} stores the numeric matrix and optional per-axis labels.
}
