% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-quant_matrix.r
\name{QuantMatrix}
\alias{QuantMatrix}
\title{Quantification Matrix Class}
\usage{
QuantMatrix(mat)
}
\arguments{
\item{mat}{A numeric-like matrix (or matrix-coercible object).}
}
\description{
\code{QuantMatrix} stores a numeric double matrix.
Axis labels should be carried by \code{dimnames(mat)}.
}
